Modified some layout details to fit rubocop and changed greater_than to greater_than_or_equal on some models. Made order in question table integer instead of float

+++ db/migrate/20210407101625_create_questions.rb
@@ -3,7 +3,7 @@ class CreateQuestions < ActiveRecord::Migration[6.1]
     create_table :questions do |t|
       t.string :title
       t.float :points
-      t.float :order
+      t.integer :order
       t.float :time_limit 
       t.integer :answer_type
       t.belongs_to :quiz
